LPS-101717 SF Rename
Not a Web module
