+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-ip-address-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Multi-Factor Authentication IP Address Implementation
+Bundle-SymbolicName: com.liferay.multi.factor.authentication.ip.address.impl
+Bundle-Version: 1.0.0
