+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-ip-address-impl/src/main/java/com/liferay/multi/factor/authentication/ip/address/internal/constants/MFAIpAddressEventTypes.java
@@ -12,7 +12,7 @@
  *
  */
 
-package com.liferay.multi.factor.authentication.ip.address.web.internal.constants;
+package com.liferay.multi.factor.authentication.ip.address.internal.constants;
 
 /**
  * @author Marta Medio
