--- modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-ip-address-web/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Multi-Factor Authentication IP Address Web
-Bundle-SymbolicName: com.liferay.multi.factor.authentication.ip.address.web
-Bundle-Version: 1.0.0
